requirements.txt
django==1.9.1
django-modeltranslation==0.11rc2
django-extensions==1.6.1
django-floppyforms==1.6.0
django-redis==4.8.0
django-dbbackup==3.2.0
django-storages==1.6.6
python-dateutil==2.4.2
pillow==5.2.0
redis==2.10.6
psycopg2==2.8.4
pyqrcode==1.1
weasyprint==54.3
pyyaml==3.11
